Don't store a copy of the device state in SoundplaneModel
The interaction between SoundplaneModel and SoundplaneDriver is currently
racy. This is a first step towards fixing that.

The problem before this patch (which still remains after it, but it is
easier to fix after this) is that SoundplaneDriver invokes deviceStateChanged
in a way that makes it possible for it to drop certain states and send
other states more than once. This is a problem, for example if
kDeviceIsTerminating is sent twice and SoundplaneModel attempts to join
its processing thread twice.